Feature: locks
Based on a FIFO queue by Maged M. Michael and Michael L. Scott which
relies on Double Word CAS:
<http://www.cs.rochester.edu/u/scott/papers/1996_PODC_queues.pdf>

Implements a lock with owner and waitlist with direct control of
fiber suspend/resume states.

Also fixes an issue in `scheduler_reschedule` which could prevent the
current fiber's context to be saved, thus be resumeable by another
thread, which would block the thread and the fiber until the
scheduler could steal some work (which could wait forever).