Fix: scheduler didn't consume its own runnables on startup
A thread main function only relies on next_runnable to start running
fibers. The problem is that next_runnable only tried to steal work
from other schedulers, preventing a scheduler from even trying to
run its own queue when it was just started, and sometimes when
rescheduling.

This is now fixed.