Locks: use a Ladan-Mozes and Shavit non blocking queue.
Instead of the Michael and Scott deque. This version seems a little
less efficient on furious lock/unlocks but allows to avoid allocating
a dummy node for each fiber (and allows to reuse a node per fiber),
which proves faster than MS-queue.

NOTE: maybe using the same technique to replace the left over node
when the queue is emptied (using a dummy node) could allow to use
the MS-queue, instead.